fix(native-host): harden browser extension integration
