+++ src/core/bridge/mbp1/pair-session.test.ts
@@ -50,7 +50,7 @@ const PROOF_LABEL = utf8ToBytes('MBP1/ticket-proof/v1')
 const INSTANCE_ID = '0d9c2b7a-4e6f-4a1b-8c3d-2e5f7a9b1c4d'
 const SERVER_GENERATION = '3c2b1a09-8f7e-4d6c-b5a4-93827160f0e1'
 const LOCAL_TOKEN = 'vector-local-token-0123456789abcdef'
-const EXTENSION_ID = 'ibpkjhgpbidfmbmomagmldcdlpbmchgi'
+const EXTENSION_ID = 'lggbokfckofcgjndaboioakcmincinpo'
 const ORIGIN = `chrome-extension://${EXTENSION_ID}`
 const INSTALLATION_ID = '5f0b6f9e-8a3d-4c5e-9b2a-7d1e4f6a8c0b'
 const PAIR_NONCE = 'vec-nonce-8f3a1c5e7b2d4a90'
