+++ src/shared/config/native-messaging-extensions.json
@@ -1,4 +1,7 @@
 {
-  "chromium": ["ibpkjhgpbidfmbmomagmldcdlpbmchgi"],
+  "chromium": [
+    "lggbokfckofcgjndaboioakcmincinpo",
+    "efcflljngohddnmfmebiamigoikmdfbf"
+  ],
   "firefox": ["motrix-extension@motrix.app"]
 }
